INTERAC-21 fix domain for Bamboo
Previously we committed the domain association file for dev-demo and demo,
and set up Jenkins to use the file for dev-demo (since dev-demo is deployed
to by Jenkins).  This meant however that when Bamboo builds the master branch
it still also uses the same domain association file for dev demo (incorrectly).
This change makes it so that the default is the domain association file for
demo is used, and Jenkins specifies the dev-demo association file explicitly
when the branch isn't master