Update the readme & add a .markdownlint.json file
As part of the Code Quality Challenge, day 1 was to update the README
of a project.  As such I updated this project's readme to be more up
to date, and remove some markdown lint errors.

To support this I also added a .markdownlint.json file to encode the
Markdown coventions we want to adopt.